[XLA:GPU] Swap dot operands in certain cases.
On H100, the `wgmma` instruction supports lhs operand in registers but not rhs operand. That means that there is any prologue for one of the sides (e.g. conversion), it's better to put it into the lhs -- as elements will be in registers after computing the prologue.

Additionally, minimal lhs non-contracting dimension size (M dimension) is 64, while for rhs smaller sizes are supported.

Because of that, after the Triton fusion is known, swap lhs and rhs ops in the following cases:

- If rhs has prologue, and lhs doesn't, and N-size (rhs non-contracting dim) is ≥64, swap.
- If M<64 and N≥64, swap.

PiperOrigin-RevId: 699151018